touyana_0624 push1

+++ Kojiro/src/main/java/com/example/Kojiro/controller/QManagementController.java
@@ -9,6 +9,7 @@ import com.example.Kojiro.entity.TestQuestionP2;
 import com.example.Kojiro.entity.question;
 import com.example.Kojiro.form.QuizManagement;
 import com.example.Kojiro.service.QManagementService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -22,11 +23,14 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Base64;
 
+
 @Controller
 public class QManagementController {
 
     private String[] successMesList={"","追加に成功","更新に成功","削除に成功"};
 
+    @Autowired
+    private HttpServletRequest request;
     private int successIndex=0;
     @Autowired
     QManagementService qManagementService;
@@ -36,6 +40,7 @@ public class QManagementController {
     private HttpSession session;
     @GetMapping("/quiz-management")
     public String menu(@RequestParam(name="keyword" ,defaultValue = "") String keyword, Model model) {
+        if(request.getSession(false)==null) return "redirect:/index";
 //        if (session.getAttribute("user")==null){//ユーザーのセッション判定
 //            return "redirect:/login-test";
 //        }
@@ -72,6 +77,7 @@ public class QManagementController {
     @GetMapping("quiz-detail/{id}/{gName}")//id指定で詳細のページを出す
     public String detail(@PathVariable("id") int id,
                          @PathVariable("gName") String gName, Model model) {
+        if(request.getSession(false)==null) return "redirect:/index";
         System.out.println(id);
         if(gName.equals("危険予測ディスカッション")){
             var result = qManagementService.findById2(id);
@@ -109,6 +115,8 @@ public class QManagementController {
     @GetMapping("quiz-update/{id}/{genre}")//id指定で更新のページを出す
     public String update1(@PathVariable("id") int id,
                           @PathVariable("genre") String gName, Model model) {
+        if(request.getSession(false)==null) return "redirect:/index";
+
         System.out.println(id);
         var genres = genresDao.findAll();
         if(gName.equals("危険予測ディスカッション")){
@@ -156,6 +164,7 @@ public class QManagementController {
 //    }
     @PostMapping("quiz-update/{id}/{genre}")
     public String update2(@PathVariable("id") int id, @PathVariable("genre") String gName, @Validated @ModelAttribute("q_management") QuizManagement update, BindingResult bindingResult, Model model){
+        if(request.getSession(false)==null) return "redirect:/index";
         if(bindingResult.hasErrors()) {//バリデーションチェック
             return "quiz-update";
         }else {
@@ -199,6 +208,7 @@ public class QManagementController {
 //    }
     @GetMapping("/quiz-delete/{id}/{genre}")//問題の削除(idを指定)
     public String delete1(@PathVariable("id") int id, @PathVariable("genre") String gName){
+        if(request.getSession(false)==null) return "redirect:/index";
         if(gName.equals("危険予測ディスカッション")){
             var result = qManagementService.delete2(id);
             successIndex=3;
@@ -235,6 +245,7 @@ public class QManagementController {
 //    }
     @PostMapping("/quiz-add")
     public String product1(@Validated @ModelAttribute("add") QuizManagement add, BindingResult bindingResult) {
+        if(request.getSession(false)==null) return "redirect:/index";
         System.out.println(add);
         if (bindingResult.hasErrors()) {
             System.out.println("バリデーション");
